+++ NEWS.md
@@ -1,8 +1,10 @@
 # mlr3hyperband (development version)
 
+* feat: Add `OptimizerAsyncSuccessiveHalving` optimizer.
+
 # mlr3hyperband 0.6.0
 
-* compatibility; Work with new bbotk 1.0.0 and mlr3tuning 1.0.0
+* compatibility: Work with new bbotk 1.0.0 and mlr3tuning 1.0.0
 
 # mlr3hyperband 0.5.0